requirements.txt
pymupdf
tqdm
scikit-learn
datasketch
groq
python-dotenv